I'd rather not describe all this

+++ src/main/kotlin/dev/deftu/componency/animations/AnimationTime.kt
@@ -1,16 +1,14 @@
 package dev.deftu.componency.animations
 
-import java.util.concurrent.TimeUnit
-
 public class AnimationTime(
     public val unit: TimeUnit,
-    public val value: Long
+    public val value: Double
 ) {
 
     public companion object {
 
         @JvmStatic
-        public val ZERO: AnimationTime = AnimationTime(TimeUnit.MILLISECONDS, 0)
+        public val ZERO: AnimationTime = AnimationTime(TimeUnit.MILLISECONDS, 0.0)
 
     }
